Move ipam tests to their own file:
